r1s: fix nailchipper quadruple crossing callouts
During the Leaping Quadruple Crossing given by one of Black Cat's
clones, players are hit by Nailchipper and given a headmarker which
indicates that they will do a mini explosion and will die if they get
hit by the next Quadruple Crossing bait. The markers always go either on
all DPS or on all support players.

Currently, when this marker goes out, cactbot incorrectly calls out
"stack on <player>" for whichever player happens to be first in the data
log. This could in theory be used to see who has the nailchippers and
processed into a decision of whether to go in or out. However, using the
term "stack" is very odd since it is *not* a stack mechanic.

Instead, check what role the target is. Based on this, alert either
"Support out" or "DPS out".

We could complicate this further by checking the role of the player and
alerting either in our out, (or some other alert such as baiting).
While this could be doable its more complicated to get correct.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>