raidboss: add r2s alert for dropping puddles
During the 2nd Alarm Pheromones, players get targeted with a head marker
indicating that they will soon have a poison puddle drop on them. The
player needs to notice this, and then take the puddle out to the edge.

Add an alert for this headmarker to warn the player when the puddle is
about to drop on them.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>