raidboss: add callout to swap or stay during XFold Blast
Currently, the callout for XFold Blast accumulates and then calls out
the chain of N/S one time near the start of the mechanic.

This works ok, but it can be easy to get lost in the middle and lose
track of whether and when you should swap.

To handle this, detect when the previous XFold Blast hits, and call out
the next platform direction. To help aid in recall if you need to swap
or stay, keep track and call out the direction with an alert level when
you need to swap, and an info level when you need to stay.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>