raidboss: prioritize north/south during Chasm of Vollok
For Zoraal Ja EX, the sword cleave callout currently prioritizes calling
the safe spot in order of:

1) inside east/west
2) the north or south corner, leaning to avoid the cleaves
3) finally the outer intercard pairs

This works, but most parties have the party stack south with the main
tank at the north. Calling the safe spot with north or south tile first
is less movement for both the maintank and party.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>