raidboss: add chasm of vollok callout configuration
The Chasm of Vollok callouts for the sword sync mechanic requires
identifying which diamonds are safe, along with avoiding a cleave that
hits the entire east or west half of the arena.

We cannot reasonably call every valid safe diamond, as this would make
the callout unusable.

Currently, the code prioritizes the inner diamonds, calling "inside
east diamond" or "inside west diamond". It also never calls the north or
south corner as safe.

My preferred strategy for this mechanic is to hover around the north or
south corner and move very slightly to adjust for the cleave into one of
the nearby safe diamonds.

Attempting to make this the default is not suitable, as many players are
likely already familiar with the existing callouts.

Introduce a configuration option to select the priority for callouts. It
can be one of:

  1. Inside Tiles

    The current behavior, prioritize the inside diamonds first.

  2. North and South Corners

    Call "North/South Corners - Lean East/West" with top priority,
    followed by "Inside North/South Diamonds - Lean East/West".

  3. North Corner

    Same as (2), but only call the north side.

  4. South Corner

    Same as (2), but only call the south side.

This gives flexibility to enable users to select the style they prefer,
without changing any default behavior.

Implement this by creating a priority lookup system. This minimizes the
code duplication, though we still end up having to check the
configuration value to handle the north/south splitting.

I opted to keep the complexity of the north/south splitting as a simpler
callout is preferable if a user knows they will always be on the north
or south corner to start.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>