raidboss: introduce helper function for stored cleaves
In m4n, we need to take a list of stored cleave directions and convert
them into cardinal safe spots. This is done in a few places, and a
future change is going to add even more.

Make StoredCleave an explicit type, and introduce a getCleaveDirs helper
function which takes the actors and stored cleave data, and returns an
array of DirectionOutputCardinal names.

This avoids duplicating the somewhat complex logic of looking up actor
data and finding the facing.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>