raidboss: add config for sidewise spark callouts
Currently, the sidewise spark callout is handled by making one initial
callout with the first two cleave safe spots, and then a big collected callout
detailing the entire sequence of safe spots.

Add a new option to have the callouts come in sequence triggered when
the previous sidewise spark abilities from each clone hit.

This option is helpful for those who find the existing trigger difficult
to process. In particular, the existing callout includes a safe spot
which has already resolved, as the very first clones begin hitting
before the final callout. This can make it somewhat confusing to process
the callout, as the directions are stale. The user must process that the
first spots have already been hit.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>