raidboss: collapse sidewise callout when possible
During the mechanic where multiple clones hit with sidewise spark, it is
often the case that we will get a pattern like the following:

  N => W => N => W => N

Here, there is a single intercard safe spot that the player could stay
for the rest of the mechanic. When this occurs, we ought to call out
just the one safe spot.

Update the getCleaveDirs function to check if all the remaining safe
directions match a single intercardinal. If they do, return just that
element.

In the R4N Sidewise Spark callout, check if we only get a single element
back, and use a specialized "<dir> => stay" callout which will indicate
that the player can stay in that single intercard for the rest of the
mechanic. Since we know the player can stay, delete the remaining stored
cleaves.

Modify the flow so that we check for this regardless of the sequence or
collected strategy.

This works for both strategies, because we have every possible remaining
cleave at the point where this trigger occurs. For the sequence
strategy, it should also detect cases where the last 4 elements are a
single spot. This is because the first element will have already been
removed after being hit.

This generally reduces the noise for many cases. Note that the sequence
strategy may still call out multiple outputs even if they are similar. I
haven't yet figured out a good way to solve this which doesn't
accidentally clear the final hit from the boss.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>