raidboss: do not "collapse" single directions in r4n
It was reported by Wexx during testing that the single-hit cleaves were
being output with intercardinal directions. This happened because
getCleaveDirs would check if a single list of ['dirN'] matched something
in the list of ['dirN', 'dirE'], and then return ['dirNE'].

It doesn't make sense to collapse anything if there is only one
direction. Fix this.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>