raidboss: don't collapse first 2 hits in collected mode
Add a mode parameter to getCleaveDirs, which can be either 'collapse' or
'keepAll'. If the mode is 'keepAll' then the collapsing logic is
skipped.

Use this to avoid collapsing the first two hits when in collected mode.
This avoids the behavior change with collected mode, so that we no
longer output an intercard at the start.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>