raidboss: show number of times hit for sequence intercards
In addition to showing the intercard position, show the total number of
times that spot will be safe.

Do this by checking the number of cleaves. If we get a collapsed cleave,
also include x2, x3, x4, or x5 in the output.

I tried a couple of ways to get the number turned into a string, but had
type errors with other methods.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>