raidboss: call a simple stay if next safe spot is the same.
Convert the sequence mode trigger for Sidewise Spark into a response,
and modify the trigger to call a simple "stay" if the next safe spot is
in the same area, and an alert with the direction otherwise.

I'm not certain if this is exactly what Wexx suggested, and it needs
some tests.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>