raidboss: new method for sequence callout
In 'sequence' mode for sidewise spark, check the previous callout when
determing where to go next. If the spots are the same, call a simple
"stay" infoText. Otherwise, call the next safe spot using an alert.

In addition, for the final collected hit from the boss, check the
previous spot and if its the same, exit early (still clearing the list
of stored cleaves if there is only one safe spot).

This ensures that we avoid a duplicate callout in sequence mode, and
stops the output as soon as we're safe.

This should hopefully meet the request from review, minimizing the
overall amount of duplicate callouts.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>