raidboss: add support for DN strategy for Ion Cluster in m4s
The standard strategy for resolving the debuffs during Ion Cluster is
the "DN" strat, which uses assigned positions for each debuff based on
the rivets in the platform.

For the blue donut and point-blank AoEs, players go to the two rivets
closest to the boss, with support going left, DPS going right.

For the purple "get hit" debuff, players go to the rivet on the very
middle of the square platform chunk.

For the near/far conal debuff, that player goes just in front of the
middle rivet closest to the player on the middle rivet.

Add a strategy option for ion cluster which can select between the
current default of no strategy and the DN strategy. To make this a bit
less verbose, convert the callout into a response and dynamically assign
the output strings based on the strategy. This could conceivably enable
additional strategies easily if other data centers use something else.

For this to work, the two sets of output strings must have support and
dps specific variants of the roundhouse and spinning conductor debuffs.

While the strategy involves use of the rivets on the square tiles, the
callouts are simplified to use "front left" or "front right" for the
aoes, and "front of middle" or "middle, behind current" for the
cone-related debuffs.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>