raidboss: add support for DN strategy during Witch Hunt in m4s
During Witch Hunt, players need to bait aoes so that no player takes the
Witch Hunt AoE twice. Currently, we callout either "Baits near (Party
Far)" or "Baits Far (Party Near)".

Users can customize these if they use a strategy which has fixed
ordering (say, Tanks, Healers, Melee, Ranged). However, the standard NA
strat is the "DN" strat which uses a flexible ordering for update, where
Tanks and Melees prioritize the close/in spots. This cannot be
implemented as-is just with custom strings.

Add a new strategy option which makes this callout work. I tried to make
a shared function it ended up being worse than just duplicating the
local lambda function used here.

In addition to modifying the alert callout for when to be in/out and
close/far, also modify the combined starting callout. Instead of calling
"close/far" for the baits, call the order of who baits, i.e.:

  "In (Tanks) => Out (Healers) => In (Melee) => Out (Ranged)"

This assumes that you understand that tanks and melee always take close
baits, and healers and ranged always take far baits. Ofcourse, the alert
at each stage also calls the close/far for you.

This strategy has one slight downside: for alerts, it will not say "Bait
Near" or "Bait Far" for DPS players, instead it says "Melee Near" or
"Ranged Far". This is because many parties may have more than 2 'ranged'
DPS jobs, and one of the ranged players may be doing 'fake melee' spots
for the DN strat. I considered adding a 'strict' version of the strategy
which would check the job and make a separate callout, but I decided
against this.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>