raidboss: add uptime sunrise sabaath callouts in m4s
Currently the sunrise sabaath callouts do not include any strategy
information. This makes the callout indicate both towers, as well as the
spots for both cannons.

The standard NA strategy is the "AutoCAD Uptime" strategy which uses the
uptime waymarks, and a priority system where support players start
looking NW and counter clockwise while DPS players look north and
clockwise.

For towers, this means DPS always take the north or east tower, while
support always take the south or west tower.

For cannons, supports will take the first cannon that fits, either NW,
SW, or rarely SE. DPS will take the first that fits from NE, SE, or
rarely SW.

The uptime waymarks also have specific bait spots on the waymarks the
side closest to the towers.

Add an uptime strategy option which uses this priority and calls the
specific tower and specific cannons to bait, and includes whether to
bait on the north/south or east/west side.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>