raidboss: split uptime spots from snake prio strategy
The current 'uptime' strategy for sunrise sabaath bakes in the uptime
positioning, when some groups use the same snake prio system for finding
the cannon, but do not use the uptime spots.

Split this strategy option so that you can select the bait positions for
uptime with or without the snake priority system.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>