raidboss: separate uptime and normal bait strings
The uptime strategy has you stand on the side of the waymark closest to
the cannon. This is distinct from the callout for pointing the cannon in
the normal strategy.

Fix this by using a separate output string to get the callout before
passing the string into yellowShort or blueShort.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>