raidboss: include safe players when calling pass debuff
After mustard bomb goes off, the tanks need to transfer a debuff to one
of the players who was not hit by kindling cauldron. Currently we just
call out "pass debuff".

We already have the data required to determine who is safe to pass the
debuff to. Filter the list of party names to and include the shortened
names of the players who its safe to pass to.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>