raidboss: pass final sidewise spark hit through getCleaveDirs
In the trigger for the final hit of sidewise spark from Wicked Thunder,
the cleave directions from stored cleaves are obtained, and then the
final hit is pushed onto the array.

This works for now, but will break when extending the logic of
getCleaveDirs to support collapsing the callout for intercardinal safe
spots.

Instead, push the boss's hit onto the stored cleave array first before
calling getCleaveDirs.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>