raidboss: avoid re-calling previous clone cleave safe spots
Currently, during sidewise spark, the clone cleaves and boss cleave are
collected. Two callouts are made.

As soon as the first two cleaves are collected, a callout is made for
the first two safe spots. Then, once the boss cleave is collected, all
safe spots are called out.

This is problematic, because players will see callouts like:

E => S
E => S => W => N => W

The actual clone hits are about 2 seconds apart. By the time the final
callout is made, the first clone will have already hit, and the second
clone will be about half a second away from hitting.

This requires players to remember that the data for the first 2 spots is
effectively useless, and that they really need to pay attention to the
3 final hits.

Instead of calling all 5 hits, just call out the last 3.

Based on the raid simulator, we are making a call while the 2nd hit is
about 0.5 to 1 seconds from hitting. We *could* refactor to delay this
call until the second clone has hit. However, this gives the player only
~2 seconds of reaction time from when we make a call like "W => N => W"
where they need to be west.

I am unsure which of these is better, but at a minimum we should stop
repeating information which makes the callout more difficult to follow.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>