dynamic starting champion's circiuits spots
There was an issue as well where the combatant data could be missing for the call (in raidemulator) unless a delay was added, and when delay was added it led to duplicate triggers, so the trigger was refactored to also collect the ids. I don't think delay is needed for the left/right now that it counts up to 5 adds, my guess is there must have been some small 0.1s lag of some add positions updating and it had captured an invalid value on something so the trigger just didn't fire.

With collection of Gleaming Fangs, I wrote a function to search through them for the matching one of the platform we are on and then that gets used to determine left/right portion of the call.

For the champion circuit mechanic, we get the player's position and on each new mechanic, the player's position is re-checked if they are still on the same platform. If the player changed their platform, the order is recalculated from previous order (this may need more testing, I think at least for the movement of 1 platform between mechs it looked okay in my 1 sample).

Then there is also simply if the players are staying on the platform that is not south, the order is changed to be relative. I do not have many logs where a non south platform is used and in logs that I have, usually the player is on their way to the south.

The one drawback right now is that the initial long string will have been out of date if the player moves platforms and there isn't a simple solution to replace that without re-triggering and reducing delay. I have opted to reduce the delay overall from 26 to the default as the individual mechanic calls take over and are based on the platform you are on at the time of the cast starting.