raidboss: add skeleton file for south horn triggers
Add a skeleton file for the Occult Crescent South Horn critical
encounters. I don't yet have the start of a timeline or any triggers,
but I did record all of the IDs for the ActorControl messages for being
"teleported" into each critical encounter. It appears to function
relatively similar to the Bozjan Southern Front.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>