raidboss: add several basic triggers for critical encounters
Add a few basic triggers for some of the critical encounters in the
Occult Crescent. This includes a few AOEs, tank busters, etc. I've
tested them while I created them on the fly. Still missing a dozen or so
basic cone and similar mechanics, but this is a good start especially
for the bosses which use these casts while during other important
mechanics.

I have not started on any of the more complex mechanics which require
figuring out where things are in the world.

I'm also open to suggestions for some of the callouts, especially the
"Sides => Front/Back" and "Front/Back => Sides" callout.

I did try to create conditional checks on whether the CE was identified,
but that didn't seem to be working properly yet.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>