encounter_tools: attempt to support CEs without breaking others
I'm not certain this works properly yet, but it appears to locate the
CEs if I merge with my occult crescent timeline file.

Instead of blindly listing all CEs with their ID, I added look up tables
for the CEs to resources. I suspect there is a more elegant way to do
this, with less boiler plate code. I also don't really know how to share
this data with the trigger files, which need inverted mappings.

This method is similar to some code written by Valarnin, but I adapted
it so that it will only apply to known valid CEs. That should prevent
breaking any other zones or encounter logic. I'm open to other methods
of integrating CE support if there are better solutions.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>