raidboss: r7s puddle baits during sinister seeds.
Sinister seeds target either the 4 farthest players or the 4 nearest
players (at random). Whoever does *not* get targetted will have puddles
appear undearneath them, followed by several more chasing puddles they
need to bait.

Although the A571 ability exists in the logfile to indicate where
puddles spawn, it does not include the player target name. This makes it
not ideal for determing who got puddles.

Instead, keep track of everyone who gets a sinister seed. Once we get 4,
then callout to bait puddles for everyone else.

This logic keeps the callout for the individual seeds players immediate
(meaning there should be minimal to no change in the timing of the
alert). The bait players will get a callout slightly later, but this is
pretty much the earliest we can call it out. Attempting to use the
player position to compare to the puddle baits is likely to be too
fragile.

To help with audio cues, I opted to downgrade the seeds to an info and
put the baits at an alert. You have less time to respond to the bait
notification, and I wanted to make sure these different roles used
different audio cues to aid in recognition of what you're supposed to
do.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>