raidboss: add syncs for Raking Scratch abilities
The Raking Scratch ID is used twice in succession, and there's little
reason to display two entries, so I tried to keep only the later entry.
Unfortunately, this resulted in some weird syncs due to the timeline
syncing to the first Raking Scratch ability, messing up the timing for
future entries.

Fix this by adding --sync-- entries for the missing Raking Scratch
abilities. Reduce their window to 1,1 so that they don't accidentally
cause a sync with the second Raking Scratch. Adjust the timeline values
to account for these entries properly.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>