util: add support for sheets with subrow_id to xivapi
The DynamicEventSet sheet contains both a row_id and a subrow_id. If we
query the sheet with "after=3" it returns all events which have a row_id
of 3 or greater *and* which have a subrow_id of greater than 0. To
properly limit the returned rows, we need to use "row:subrow" syntax
with the after field.

Extend the logic to obtain the subrow_id and add it to query, ensuring
that we can safely iterate over the DynamicEventSet sheet in a following
change.

Currently appending ":0" to regular after=N "works", this is likely not
intentional and should not be relied upon. Instead, only append the
subrow indicator if the subrow_id exists.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>