lancelight and more rune axe fixes
use of the boolean led to more issues, all seem to be resolved using a counter instead.

lancelight was missing some calls later due to Magitaur vs Luminous Lance as the source... switching to the A256 cleans it up as it is just cast by Luminous Lance and it isn't two different spells. A256 was only visible in a network log and not seen from fflogs.