boss 1 second tower spawns, earlier gargoyle safe spot
The earlier call still needs to be tested in game and so I am leaving the startsUsingExtra trigger for now to update the data .

While the raidemulator will have the data, I doubt in game it works since the raidemulator would show the data before the thing would be spawned when using startsUsing timing.

This would also apply to the Icicle callout in the 3rd boss. I think the log actually has the data about 2 seconds after the cast ends, but it is visible in game at cast end.