add From Times Bygone timeline
The fight has a long loop block, in it there are some predictions that can be made, but cannot predict the double death ray/big burst or if there is one of each. Can predict the Steelstrike ones though which this attempts to do, but adding these predictor branches rolls up close to the next CE's start time.