add Assassin's Dagger pattern call
This also adds configuator for BAP Daggers. Usually cactbot tries not to use a specific strategy, but this might be a case where BAP Daggers is so widely used it should be the default.

Currently just names the pattern, but with better Assassin's Dagger counting triggers, we can call when to move into square/update axeblow/lanceblow when safe to dodge.