raidboss: Arkveld Extreme timeline
Add a rough timeline for Arkveld (Extreme). I did actually have a log
file that went long enough to hit enrage, so I am confident in the end
of the fight. I opted to avoid trying to branch for the spots where it
alternates between the red and white dashes, as this was the only
branching RNG in the fight from what I could tell.

This needs live testing, as I only correlated against log files so far.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>