remove synthetic CEs for forked tower due to encounter issues
Trying to include the CE IDs for the DirectorUpdate lines within the
forked tower is problematic. Each area gets its own DirectorUpdate upon
completion, but the main encounters already have seal/unseal lines to
track the encounter. Trying to treat these as CEs results in a bunch of
0-length encounters, and this results in confusion when checking
timelines with the log tools.

Theoretically we could try to use the IDs for tracking the central
passages and the lockwards. Unfortunately, this is problematic because
the DirectorUpdate happens *before* the unseal messages from the game.

For now, just remove these synthetic IDs and leave a comment detailing
the IDs for the future. If someone wants to try and fix the logtools to
better track this somehow they will have the information to help with
this.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>