raidboss: add strategy for Dead Stars Vengeful Direction
The Vengeful Direction attack during Dead Stars in the Forked Tower:
Blood is sometimes called out using standardized Waymarks. This is true
for several discords who organize the tower, FOE, ABBA, and CAFE.

Add a strategy option to allow configuring the callout to use the
waymark, the direction, or both. If for any reason the callout direction
is not one of the expected waymark direction, we fall back to calling
the direction along with '???'.

This option makes it easier to align with standard strategies based
around the waymarks.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>