raidboss: fix incorrect "Follow Dash => Get Behind" trigger
The trigger for sweeping charge was incorrectly placed on both the start
attack with the dash, *and* the followup sweep, causing cactbot to
trigger twice with the same text. Fix this by moving the later A4C5
trigger to just be the getBehind callout for Soul Sweep. This is the
same as how Menacing Charge is handled.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>