raidboss: fix OC CE timelines if already in combat
Most of the CE timelines have an InCombat sync line to ensure that the
timeline matches up with what we expect. In some circumstances, players
may accidentally already be in combat when the CE starts. If this
occurs, the timeline will never sync up, as the InCombat sync will not
be triggered. This results in broken timelines. In many cases the
timeline still "appears" but becomes wildly inaccurate.

Fix this by increasing the window for the first StartsUsing cast which
is used to re-sync the timeline in case of a backup. Ensure this window
is large enough to fully cover the 100 seconds skipped at the start of
entering a CE.