raidemulator: support critical encounters
The raid emulator does not use the log tools to split network log files
for encounter data. Currently this means that critical encounters in OC
and other zones do not get separated properly. This is problematic as it
means you cannot use the emulator to easily test such encounters.
Additionally, this can result in extremely long encounters which take a
long time to parse. In my case, this resulted in Firefox repeatedly
complaining about the thread taking too long and trying to stop it.

Add support for the CE start and end to the EmulatorCommon.ts regex
list to allow splitting the critical encounters properly within the
emulator.